requirements.txt
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.2
Flask_WTF==1.0.2
Flask_Mail==0.9.1
WTForms==3.0.1
itsdangerous==2.1.2
SQLAlchemy==1.4.35
bcrypt==3.2.0
python-dotenv==0.19.1